sync: export from internal SnapRHI (d9a8309b)
Automated sync from github.sc-corp.net/Snapchat/SnapRHI-Internal
Internal commit: d9a8309b572d0051af419aa0643c59a8f06a679e
Changed files: 50
Insertions: +988, Deletions: -523

+++ src/backend/opengl/src/BlitCmdPerformer.cpp
@@ -555,8 +555,9 @@ void copy(snap::rhi::backend::opengl::Device* device,
     };
     for (uint32_t i = 0; i < cmd.infoCount; ++i) {
         const auto& copyInfo = cmd.infos[i];
-        dstBuffer->uploadData(
-            copyInfo.dstOffset, std::span<const std::byte>{ptr + copyInfo.srcOffset, copyInfo.size}, dc);
+        dstBuffer->uploadData(copyInfo.dstOffset,
+                              std::span<const std::byte>{ptr + copyInfo.srcOffset, static_cast<size_t>(copyInfo.size)},
+                              dc);
     }
 }
 } // unnamed namespace
